+++ lib/asset/algo_validation.go
@@ -3,6 +3,7 @@ package asset
 import (
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 	is "github.com/go-ozzo/ozzo-validation/v4/is"
+	"github.com/owkin/orchestrator/lib/errors"
 )
 
 // Validate returns an error if the new algo is not valid:
@@ -16,5 +17,51 @@ func (a *NewAlgo) Validate() error {
 		validation.Field(&a.Algorithm, validation.Required),
 		validation.Field(&a.Metadata, validation.By(validateMetadata)),
 		validation.Field(&a.NewPermissions, validation.Required),
+		validation.Field(&a.Inputs, validation.By(validateInputs)),
+		validation.Field(&a.Outputs, validation.By(validateOutputs)),
 	)
+}
+
+func validateInputs(input interface{}) error {
+	algoInputs, ok := input.(map[string]*AlgoInput)
+	if !ok {
+		return errors.NewInvalidAsset("inputs is not a proper map")
+	}
+
+	for name, input := range algoInputs {
+		err := validation.Validate(name, validation.Required, validation.Length(1, 100))
+		if err != nil {
+			return err
+		}
+
+		err = validation.ValidateStruct(input,
+			validation.Field(&input.Kind, validation.In(AssetKind_ASSET_MODEL, AssetKind_ASSET_DATA_SAMPLE, AssetKind_ASSET_DATA_MANAGER)))
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func validateOutputs(input interface{}) error {
+	algoOutputs, ok := input.(map[string]*AlgoOutput)
+	if !ok {
+		return errors.NewInvalidAsset("outputs is not a proper map")
+	}
+
+	for name, output := range algoOutputs {
+		err := validation.Validate(name, validation.Required, validation.Length(1, 100))
+		if err != nil {
+			return err
+		}
+
+		err = validation.ValidateStruct(output,
+			validation.Field(&output.Kind, validation.In(AssetKind_ASSET_MODEL, AssetKind_ASSET_PERFORMANCE)))
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }