+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add a new `ALGO_PREDICT` algo category (#693)
 
+### Changed
+- Validate algo inputs and outputs (#699)
+
 ## [0.10.0] - 2022-05-03
 
 ### Changed